Fix bug involving admins impersonating teachers
