+++ services/QuillLMS/spec/factories/schools_admins.rb
@@ -19,6 +19,6 @@
 FactoryBot.define do
   factory :schools_admins do
     school { create(:school) }
-    user { create(:user) }
+    user { create(:teacher) }
   end
 end
