+++ services/QuillLMS/spec/support/system.rb
@@ -4,7 +4,7 @@ RSpec.configure do |config|
   Capybara.register_driver :local_selenium_chrome_headless do |app|
     options = Selenium::WebDriver::Chrome::Options.new(
       args: [
-        'headless',
+        # 'headless',
         'window-size=1920x1280'
       ]
     )
